Add validation_whitelist and use pre_validate_filter instead of javascript in confirmation
I was using javascript with hidden fields to assing 1 or 0 to the value of the confirmation
field type.
Now I'm using validation_whitelist to check using "is_valid" method the $field_value. The
only valid option is '1'.
I'm using "pre_validate_filter" when the checkbox is not checked, $field_value is empty and
the field is not rendered in the profile loop because "field_has_data" is false. I put a '0'
in $field_value so "field_has_data" will be true and the field will be rendered always.