Rewrites "is_valid" method for taxonomy field type
I've rewritten the validation method for taxonomy field type. In "validation_whitelist"
I have the name of taxonomy selected by the admin. So I'm replacing this value with the
ids of the tags and call parent::is_valid.
So the normal "is_valid" method from buddypress will check the id selected with all the
ids of the tags from the taxonomy